fix(nvim): stop assuming opened file has final eol
The `vim-sleuth` extension was also setting the `&eol` to true, even
when the opened file didn't have a final eol character. Swapping it with
a different extension seems to have resolved the issue. I'm okay with no
longer auto detecting indentation settings, in favour of setting
editorconfig settings per project.