fix(i3status-rs): remove github block
Setting an OAuth token in the config file makes it annoying to commit
changes to it because of my `ripsecrets` pre-commit hook. Setting an
environment variable is too much of a hassle because of my lack of
understanding of starting an X session from scratch/first principles.
So, I decided to stop trying deal witht eh token disappearing and stuff.