feat(tmux): change prefix to `C-]`
I wanted to use a keybinding that won't interfere with common vim/emacs
key bindings. `C-Space` didn't seem to work due to it matching `C-@` on
Wezterm/Alacritty. Terminal escape/control codes are complicated -_-

`C-]` conflicts with a keybinding in emacs/bash, which I'm fine with for
now. The keys are close enough to the home row for the time being.